requirements.txt
boto3==1.42.62
botocore==1.42.62
certifi==2026.2.25
charset-normalizer==3.4.5
dnspython==2.8.0
idna==3.11
jmespath==1.1.0
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
